Corrected path to MacOS Freeplane directory
The Freeplane Add-on plugin was not working on MacOS because the plugin
was adding 'Contents/Java' to the path. 'Contents/app' is the correct
addition. Maybe it is better to leave this addition away and put the
full Freeplane directory path for MacOS  in the README.
In addition './gradlew packageAddon' works under JVM 11 and 17 as well.